added rvalue references to some functions

+++ Game.cpp
@@ -8,7 +8,7 @@ const float Game::WIDTH = 320.f;
 const float Game::HEIGHT = 240.f;
 
 void Game::setup() {
-	window.create(VideoMode(WIDTH, HEIGHT), "Rouge Like", Style::Default);
+	window.create(VideoMode(static_cast<unsigned int>(WIDTH), static_cast<unsigned int>(HEIGHT)), "Rouge Like", Style::Default);
 	window.setFramerateLimit(120);
 	view.setSize(WIDTH, HEIGHT);
 	cullingPoint.setSize(Vector2f(WIDTH, HEIGHT));
